fix: move prep_request_with_timeout into utils
