+++ src/cache/requests/scalar/get.rs
@@ -1,7 +1,7 @@
 use crate::cache::requests::MomentoRequest;
 use crate::cache_client::CacheClient;
-use crate::simple_cache_client::prep_request_with_timeout;
 use crate::utils::parse_string;
+use crate::utils::prep_request_with_timeout;
 use crate::MomentoErrorCode;
 use crate::{IntoBytes, MomentoError, MomentoResult};
 use momento_protos::cache_client::ECacheResult;
