+++ src/cache/requests/set/set_add_elements.rs
@@ -2,7 +2,7 @@ use momento_protos::cache_client::SetUnionRequest;
 
 use crate::cache::requests::MomentoRequest;
 use crate::cache_client::CacheClient;
-use crate::simple_cache_client::prep_request_with_timeout;
+use crate::utils::prep_request_with_timeout;
 use crate::{CollectionTtl, IntoBytes, MomentoResult};
 
 /// Request to add elements to the given set. Creates the set if it does not exist.
