+++ src/cache/requests/sorted_set/sorted_set_fetch_by_rank.rs
@@ -3,7 +3,7 @@ use momento_protos::cache_client::{SortedSetFetchRequest, Unbounded};
 
 use crate::cache::requests::sorted_set::sorted_set_fetch_response::SortedSetFetch;
 use crate::cache::requests::MomentoRequest;
-use crate::simple_cache_client::prep_request_with_timeout;
+use crate::utils::prep_request_with_timeout;
 use crate::{CacheClient, IntoBytes, MomentoResult};
 
 #[repr(i32)]
